refactor: 패키지 구조 변경  (bugpigg/issue-tracker#1)
