+++ BE/src/main/java/com/team09/issue_tracker/comment/Emogi.java
@@ -1,9 +1,12 @@
-package com.team09.issue_tracker.comment.domain;
+package com.team09.issue_tracker.comment;
 
+import com.team09.issue_tracker.comment.Comment;
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
+import javax.persistence.JoinColumn;
+import javax.persistence.ManyToOne;
 
 @Entity
 public class Emogi {
@@ -13,5 +16,9 @@ public class Emogi {
 	private Long id;
 
 	private String unicodeValue;
+
+	@ManyToOne
+	@JoinColumn(name = "comment_id")
+	private Comment comment;
 
 }
