+++ BE/src/main/java/com/team09/issue_tracker/comment/EmogiController.java
@@ -1,5 +1,6 @@
-package com.team09.issue_tracker.emogi.controller;
+package com.team09.issue_tracker.comment;
 
+import com.team09.issue_tracker.comment.dto.EmogiCreateRequestDto;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
