--- BE/src/main/java/com/team09/issue_tracker/emogi/domain/Emogi.java
@@ -1,24 +1,0 @@
-package com.team09.issue_tracker.emogi.domain;
-
-import com.team09.issue_tracker.comment.domain.Comment;
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import javax.persistence.JoinColumn;
-import javax.persistence.ManyToOne;
-
-@Entity
-public class Emogi {
-
-	@Id
-	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Long id;
-
-	private String unicodeValue;
-
-	@ManyToOne
-	@JoinColumn(name = "comment_id")
-	private Comment comment;
-
-}
