+++ BE/src/main/java/com/team09/issue_tracker/member/MemberController.java
@@ -1,4 +1,4 @@
-package com.team09.issue_tracker.member.controller;
+package com.team09.issue_tracker.member;
 
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
