+++ BE/src/main/java/com/team09/issue_tracker/milestone/dto/MilestoneRequestDto.java
@@ -1,4 +1,4 @@
-package com.team09.issue_tracker.milestone;
+package com.team09.issue_tracker.milestone.dto;
 
 import java.time.LocalDateTime;
 import lombok.Getter;
